requirements.txt
pyserial>=3.5
pymodbus>=3.6,<4.0
matplotlib>=3.8
pandas>=2.2
openpyxl>=3.1
PyYAML>=6.0
pyinstaller>=6.0
